Fix another compiler warning on Win32
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@3314 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fl_BlockLayout.cpp
@@ -4755,15 +4755,15 @@ void    fl_BlockLayout::StopList(void)
 			{
 				pStyle->getProperty("margin-left", szAlign);
 				pStyle->getProperty("text-indent", szIndent);
-				fAlign = atof(szAlign);
+				fAlign = (float)atof(szAlign);
 				fAlign *= level;
 				sprintf(align, "%fin", fAlign);
 				sprintf(indent, "%s", szIndent);
 			}
 			else
 			{
-				fAlign = 0.25 * level;
-				fIndent = -0.25;
+				fAlign = (float)(0.25 * level);
+				fIndent = (float)-0.25;
 				sprintf(align, "%fin", fAlign);
 				sprintf(indent, "%fin", fIndent);
 			}
